+++ cognite/client/data_classes/__init__.py
@@ -109,6 +109,7 @@ from cognite.client.data_classes.data_sets import (  # isort: skip
 
 from cognite.client.data_classes.transformation_notifications import TransformationNotification, TransformationNotificationList
 from cognite.client.data_classes.transformation_schedules import TransformationSchedule, TransformationScheduleList, TransformationScheduleUpdate
+from cognite.client.data_classes.transformation_schema import TransformationSchemaColumn, TransformationSchemaColumnList
 from cognite.client.data_classes.transformations import Transformation, TransformationList, TransformationUpdate, TransformationDestination, TransformationJobBlockade, OidcCredentials
 
 from cognite.client.data_classes.shared import (  # isort: skip